Refactor component and file names
